Completed refactoring
- Seperated init function
- Seperated loader (previously importer)
- Seperated Account class to models
- Seperated allocation
- Seperated cashflow
- Converted allocation function to class
- Converted cashflow funcation to class
- Updated imports in test_core
