+++ finpack/core/exceptions.py
@@ -1,0 +1,15 @@
+"""Package specific exceptions.
+"""
+__copyright__ = "Copyright (C) 2021 Matt Ferreira"
+
+
+class AccountError(Exception):
+    """Account class exception"""
+
+    pass
+
+
+class DataError(Exception):
+    """Data not provided or available as expected."""
+
+    pass
