+++ finpack/reports/allocation.py
@@ -1,0 +1,13 @@
+"""Class for building custom allocation.
+"""
+__copyright__ = "Copyright (C) 2021 Matt Ferreira"
+
+
+class Allocation:
+    """Build asset and liability allocation"""
+
+    # TODO: GH-#5 Build allocation class
+    pass
+
+
+w
