+++ finpack/reports/cashflow.py
@@ -1,0 +1,10 @@
+"""Class for building custom cashflow statements.
+"""
+__copyright__ = "Copyright (C) 2021 Matt Ferreira"
+
+
+class CashFlow:
+    """Build cashflow statement"""
+
+    # TODO: GH-#4 - Build cashflow statement class
+    pass
